src/test/java/leetcode/Solution00013Test.java
package leetcode;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Solution00013Test {

    @ParameterizedTest
    @MethodSource("casesProvider")
    void romanToInt(String input, int expected) {
        assertEquals(expected, new Solution00013().romanToInt(input));
    }

    static Stream<Arguments> casesProvider() {
        return Stream.of(
                Arguments.of("III", 3),
                Arguments.of("IV", 4),
                Arguments.of("IX", 9),
                Arguments.of("LVIII", 58),
                Arguments.of("MCMXCIV", 1994)
        );
    }

}